refactor: renames vendor
